Add tests and fixtures
